1731 - Made function generic and moved it to polars utils cleaning utils
