+++ CHANGELOG.md
@@ -14,7 +14,7 @@ All notable changes to this project will be documented in this file.
 
 - Added placeholders to slv prepare job, a utils script for prep with placeholders and tests placeholders.
 
-- Added merge_legacy_job_roles function to clean_ascwds_workplace job.
+- Added merge_job_role_columns function to polars_utils > cleaning_utils and called it in clean_ascwds_workplace job.
 
 ### Changed
 - Pull clean workplace columns into merge job within SLV pipeline.
